Fix a crash related to the model-initialization failing after a device-limit reached (#2187)

+++ src/models/user.h
@@ -38,7 +38,7 @@ class User final : public QObject {
 
   const QString& email() const { return m_email; }
 
-  int maxDevices() const { return (int)m_maxDevices; }
+  int maxDevices() const { return m_maxDevices; }
 
   bool subscriptionNeeded() const { return m_subscriptionNeeded; }
